Solang integration (#60)
This request adds
* Flipper contract, written in Solidity
* JavaScript test-case which uses it
* Solang dependency

Also, this request improves initialization procedure. The idea is that a user needs to have only Cargo, Yarn and Docker/Podman on his/her machine in order to run the testing bench. In case explicit paths are not provided, Solang and Substrate images are pulled and used. WABT is pulled if we can't find it in user's $PATH.